+++ source/ApplicationRoot.cpp
@@ -1,16 +1,19 @@
 #include "ApplicationRoot.hpp"
 #include <stdexcept>
+#include "GUI\GuiEngine.hpp"
+
+//-----------------------------------------------------------------
+// Defines
+//-----------------------------------------------------------------
+#define GUI_ENGINE (GuiEngine::GetSingleton())
 
 ApplicationRoot::ApplicationRoot()
 {
-	_window = nullptr;
 }
 
 
 ApplicationRoot::~ApplicationRoot()
 {
-	SDL_DestroyTexture(renderTex);
-	SDL_DestroyRenderer(sdlRenderer);
 }
 
 
@@ -25,42 +28,25 @@ void ApplicationRoot::end()
 {
 	TTF_CloseFont(m_ConsoleFontRegularPtr);
 	TTF_CloseFont(m_ConsoleFontBoldPtr);
-	delete[] pixels;
 	delete m_Renderer;
 	m_Renderer = nullptr;
-	delete guiFunctions;
-	guiFunctions = nullptr;
-	delete ofWnd;
-	ofWnd = nullptr;
 	delete imgExp;
 	imgExp = nullptr;
+	delete m_BmpResultPtr;
+	m_BmpResultPtr = nullptr;
+	delete m_BmpTexturePtr;
+	m_BmpTexturePtr = nullptr;
 }
 
 void ApplicationRoot::initSystems()
 {
-	//Initialize SDL
-	SDL_Init(SDL_INIT_EVERYTHING);
-
-	TTF_Init();
-	atexit(TTF_Quit);
+	GUI_ENGINE->Init();
 
 	ilInit();
 	iluInit();
 	ilutRenderer(ILUT_DIRECT3D9);
 
-	_window = SDL_CreateWindow("LightWhat - Robert Lindner", SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED
-		, 1900, 980, 
-		0);
-
-	SDL_GetWindowSize(_window, &_screenWidth, &_screenHeight);
-	sdlRenderer = SDL_CreateRenderer(_window, -1, 0);
-	renderTex = SDL_CreateTexture(sdlRenderer,
-		SDL_PIXELFORMAT_ARGB8888, SDL_TEXTUREACCESS_STATIC, _screenWidth, _screenHeight);
-	pixels = new Uint32[_screenWidth * _screenHeight];
-	memset(pixels, 30, _screenWidth * _screenHeight * sizeof(Uint32));
-
 	m_Renderer = new LWRenderer();
-	guiFunctions = new winmain(_screenWidth, _screenHeight); 
 	imgExp = new ImageExporter();
 
 
@@ -84,12 +70,15 @@ void ApplicationRoot::initSystems()
 		m_Renderer->init((exeDirectory + string("Resources/sceneDescriptionFiles/sceneCornell.lwtf")).c_str());
 		m_ResolutionX = m_Renderer->m_Settings.resolutionX;
 		m_ResolutionY = m_Renderer->m_Settings.resolutionY;
-		m_ImagePosX = _screenWidth - (m_ResolutionX + 50);
+		m_ImagePosX = GUI_ENGINE->GetWidth() - (m_ResolutionX + 50);
 		m_ImagePosY = 50;
+		m_BmpResultPtr = new Bitmap(m_ResolutionX, m_ResolutionY);
+
 		m_TexResolutionX = 400;
 		m_TexResolutionY = 400;
 		m_TexPosX = 50;
-		m_TexPosY = _screenHeight - (m_TexResolutionX + 50);
+		m_TexPosY = GUI_ENGINE->GetHeight() - (m_TexResolutionX + 50);
+		m_BmpTexturePtr = new Bitmap(m_TexResolutionX, m_TexResolutionY);
 
 		m_ConsoleFontRegularPtr = TTF_OpenFont((exeDirectory + string("Resources/fonts/Inconsolata-Regular.ttf")).c_str(), 20);
 		m_ConsoleFontBoldPtr = TTF_OpenFont((exeDirectory + string("Resources/fonts/Inconsolata-Bold.ttf")).c_str(), 20);
@@ -110,30 +99,33 @@ void ApplicationRoot::functionLoop()
 {
 	while(exit == false)
 	{
+		//Handle user input
+		processInput();
 		switch (m_Renderer->m_RenderStatus)
 		{
-		case LW_SETUP:
-			processInput();
-			drawImage();
-			break;
-		case LW_PREPROCESS:
-			processInput();
-			drawImage();
-			break;
 		case LW_RUNNING:
-			processInput();
+			//Render a few samples
 			updateImage();
-			drawImage();
-			break;
-		case LW_PAUSED:
-			processInput();
-			drawImage();
-			break;
-		case LW_COMPLETE:
-			processInput();
-			drawImage();
 			break;
 		}
+		//Draw Scene Info
+		SDL_Color color = { 255, 255, 255, 255 };
+		int x = m_ImagePosX, dy = 30, fSize = 20, y = m_ImagePosY - 30;
+		GUI_ENGINE->DrawString(m_Renderer->m_Status, m_ConsoleFontRegularPtr, color, fSize, x, y);
+		y = m_ImagePosY + m_ResolutionY + dy;
+		for (size_t i = 0; i < sceneInfo.size(); i++){
+			GUI_ENGINE->DrawString(sceneInfo[i], m_ConsoleFontRegularPtr, color, fSize, x, y);
+			y += dy;
+		}
+		//Draw Texture Bitmap
+		if (m_DrawTexture)
+		{
+			GUI_ENGINE->DrawBitmap(m_TexPosX, m_TexPosY, m_BmpTexturePtr);
+		}
+		//Draw Rendered image
+		GUI_ENGINE->DrawBitmap(m_ImagePosX, m_ImagePosY, m_BmpResultPtr);
+		//Make SDL render the GUI
+		GUI_ENGINE->Paint();
 	}
 }
 
@@ -159,7 +151,7 @@ void ApplicationRoot::processInput()
 				m_Renderer->preprocess();
 				break;
 			case SDLK_l:
-				fileName = guiFunctions->getFileName(ofWnd);
+				fileName = GUI_ENGINE->GetWindow()->GetFileName();
 				if ((fileName==string("nofile")))
 				{
 					cout << "loading file canceled " << endl;
@@ -194,7 +186,7 @@ void ApplicationRoot::processInput()
 				cout << "scene cleared! " << endl;
 				break;
 			case SDLK_s:
-				fileName = guiFunctions->saveFileName(ofWnd); 
+				fileName = GUI_ENGINE->GetWindow()->SaveFileName();
 				if ((fileName == string("nofile")))
 				{
 					cout << "saving file canceled " << endl;
@@ -217,7 +209,11 @@ void ApplicationRoot::processInput()
 				}
 				break;
 			case SDLK_d:
-				drawTexture();
+				m_DrawTexture = !m_DrawTexture;
+				if (m_DrawTexture)
+				{
+					drawTexture();
+				}
 				break;
 			default:
 				break;
@@ -236,7 +232,7 @@ void ApplicationRoot::updateImage()
 		{
 			for (int j = 0; j < m_ResolutionY; j++)
 			{
-				setPixel(i + m_ImagePosX, j + m_ImagePosY, m_Renderer->m_TonemappedImage->getRGB(i, j));
+				m_BmpResultPtr->SetPixel(i, j, m_Renderer->m_TonemappedImage->getRGB(i, j));
 			}
 		}
 	}
@@ -246,13 +242,11 @@ void ApplicationRoot::drawTexture(){
 	if (m_Renderer->m_Scene->textures.size() >0)
 	{
 		Texture tex = m_Renderer->m_Scene->textures[dispTexIdx];
+		for (int i = 0; i < m_TexResolutionX; i++)
 		{
-			for (int i = 0; i < m_TexResolutionX; i++)
+			for (int j = 0; j < m_TexResolutionY; j++)
 			{
-				for (int j = 0; j < m_TexResolutionY; j++)
-				{
-					setPixel(i + m_TexPosX, j + m_TexPosY, tex.getRGB(((float)i / (float)m_TexResolutionX), ((float)j / (float)m_TexResolutionY)));
-				}
+				m_BmpTexturePtr->SetPixel(i, j, tex.getRGB(((float)i / (float)m_TexResolutionX), ((float)j / (float)m_TexResolutionY)));
 			}
 		}
 	}
@@ -276,69 +270,8 @@ void ApplicationRoot::drawTexture(){
 				float x = (float)i / (float)m_TexResolutionX;
 				float y = (float)j / (float)m_TexResolutionY;
 				colRGB thisPixel = tex.getRGB(x, y);
-				setPixel(i + m_TexPosX, j + m_TexPosY, thisPixel);
+				m_BmpTexturePtr->SetPixel(i, j, thisPixel);
 			}
 		}
-	}
-}
-
-void ApplicationRoot::renderText(const std::string &message, TTF_Font *daFont,
-	SDL_Color color, int fontSize, int posX, int posY)
-{
-	//We need to first render to a surface as that's what TTF_RenderText
-	//returns, then load that surface into a texture
-	SDL_Surface *surf = TTF_RenderText_Blended(daFont, message.c_str(), color);
-	if (surf == nullptr){
-	}
-	SDL_Texture *texture = SDL_CreateTextureFromSurface(sdlRenderer, surf);
-	if (texture == nullptr){
-	}
-	//Clean up the surface and font
-	SDL_FreeSurface(surf);
-
-	SDL_Rect texture_rect;
-	texture_rect.x = posX;  //the x coordinate
-	texture_rect.y = posY; // the y coordinate
-	texture_rect.w = message.length()*fontSize/2; //the width of the texture
-	texture_rect.h = fontSize;
-
-	SDL_RenderCopy(sdlRenderer, texture, NULL, &texture_rect);
-
-	SDL_DestroyTexture(texture);
-}
-
-void ApplicationRoot::displaySceneInfo(){
-	SDL_Color color = { 255, 255, 255, 255 };
-	int x = m_ImagePosX, dy = 30, fSize = 20, y = m_ImagePosY - 30;
-	renderText(m_Renderer->m_Status, m_ConsoleFontRegularPtr, color, fSize, x, y);
-	y = m_ImagePosY + m_ResolutionY + dy;
-	for (size_t i = 0; i < sceneInfo.size(); i++){
-		renderText(sceneInfo[i], m_ConsoleFontRegularPtr, color, fSize, x, y);
-		y += dy;
-	}
-}
-
-void ApplicationRoot::drawImage()
-{
-	SDL_UpdateTexture(renderTex, NULL, pixels, _screenWidth *sizeof(Uint32));
-
-	SDL_RenderClear(sdlRenderer);
-	SDL_RenderCopy(sdlRenderer, renderTex, NULL, NULL);
-
-	displaySceneInfo();
-
-	SDL_RenderPresent(sdlRenderer);
-
-}
-
-void ApplicationRoot::setPixel(int x, int y, colRGB col)
-{
-	if (y >= 0 && y < _screenHeight && x >= 0 && x < _screenWidth)
-	{
-		int red = (int)(min(powf(col.red, (1.f / 2.2f)), 1.f) * 255) << 16;
-		int green = (int)(min(powf(col.green, (1.f / 2.2f)), 1.f) * 255) << 8;
-		int blue = (int)(min(powf(col.blue, (1.f / 2.2f)), 1.f) * 255);
-		Uint32 uint = Uint32(red + green + blue);
-		pixels[y * _screenWidth + x] = uint;
 	}
 }